+++ src/engine/tuning.h
@@ -10,20 +10,20 @@ typedef struct {
   char altLetter;
   char altAccidental[4];
   int octave;
-  double cents;
-  double frequency;
+  float cents;
+  float frequency;
 
 } Note;
 
 
 
 // given a frequency, find the closest note in equal temperament
-Note Tuning12TET_find(double freq, double pitchStandard, double centsOffset, int transpose);
+Note Tuning12TET_find(float freq, float pitchStandard, float centsOffset, int transpose);
 
 // find the nearest note from a list of pitches (in cents)
-Note Tuning12TET_findNearest(double freq, double* notesInCents, int notesLength, double pitchStandard);
+Note Tuning12TET_findNearest(float freq, float* notesInCents, int notesLength, float pitchStandard);
 
-double Tuning12TET_freqToCents(double freq, double pitchStandard);
-double Tuning12TET_centsToFreq(double cents, double pitchStandard);
-double Tuning12TET_centsMoveToOctave(double cents, int octave);
-Note Tuning12TET_centsToNote(double cents, double pitchStandard, double centsOffset, int transpose);
+float Tuning12TET_freqToCents(float freq, float pitchStandard);
+float Tuning12TET_centsToFreq(float cents, float pitchStandard);
+float Tuning12TET_centsMoveToOctave(float cents, int octave);
+Note Tuning12TET_centsToNote(float cents, float pitchStandard, float centsOffset, int transpose);
